fix(ci): python coverage shards — align hosted coverage payload and monitor test doubles
