+++ tests/unit/runtime/test_hosted_validation_delegate_part_002.py
@@ -1,0 +1,776 @@
+"""Hosted validation delegation edge-case tests."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import httpx
+import pytest
+
+from awf.profiles.models import WorkspaceProfile
+from awf.runtime.hosted_delegation import (
+    HostedDelegationProtocolError,
+    HostedValidationDelegate,
+)
+from tests.unit.runtime.test_hosted_validation_delegate import _config
+
+
+@pytest.mark.unit
+async def test_hosted_validation_sanitizes_remote_phase_before_artifact_write(
+    tmp_path: Path,
+) -> None:
+    malicious_phase = "escape/../../owned"
+    workspace_artifacts = tmp_path / "ws_hosted"
+    (workspace_artifacts / "01_escape").mkdir(parents=True)
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 0,
+                            "duration_seconds": 0.2,
+                            "stdout": "out\n",
+                            "stderr": "err\n",
+                            "phase": malicious_phase,
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+        )
+
+    command = result.commands[0]
+    assert command.phase == malicious_phase
+    assert command.stdout_path.name == "01_escape_owned.stdout"
+    assert command.stderr_path.name == "01_escape_owned.stderr"
+    assert command.stdout_path.resolve().is_relative_to(workspace_artifacts.resolve())
+    assert command.stderr_path.resolve().is_relative_to(workspace_artifacts.resolve())
+    assert command.stdout_path.read_text(encoding="utf-8") == "out\n"
+    assert command.stderr_path.read_text(encoding="utf-8") == "err\n"
+    assert not (tmp_path / "owned.stdout").exists()
+    assert not (tmp_path / "owned.stderr").exists()
+
+
+@pytest.mark.unit
+async def test_hosted_validation_polls_running_before_terminal_failure(
+    tmp_path: Path,
+) -> None:
+    poll_count = 0
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        nonlocal poll_count
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            poll_count += 1
+            if poll_count == 1:
+                return httpx.Response(
+                    200,
+                    json={
+                        "operation_id": "val_1",
+                        "workspace_id": "ws_hosted",
+                        "state": "running",
+                    },
+                )
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "failed",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 1,
+                            "duration_seconds": 0.2,
+                            "stdout": "",
+                            "stderr": "failed\n",
+                            "phase": "validate",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+        )
+
+    assert poll_count == 2
+    assert len(result.commands) == 1
+    assert result.commands[0].returncode == 1
+    assert result.commands[0].stderr_path.read_text(encoding="utf-8") == "failed\n"
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    ("terminal_payload", "match"),
+    [
+        (
+            {
+                "operation_id": "val_1",
+                "workspace_id": "ws_hosted",
+                "state": "paused",
+            },
+            "unknown state",
+        ),
+        (
+            {
+                "operation_id": "val_1",
+                "workspace_id": "ws_hosted",
+            },
+            "missing state",
+        ),
+    ],
+)
+async def test_hosted_validation_rejects_unknown_or_missing_operation_state(
+    tmp_path: Path,
+    terminal_payload: dict[str, object],
+    match: str,
+) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(200, json=terminal_payload)
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match=match):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    ("state", "expected_returncode", "expected_reason_code"),
+    [
+        ("failed", 1, "HOSTED_VALIDATION_FAILED"),
+        ("cancelled", 130, "HOSTED_VALIDATION_CANCELLED"),
+        ("timed_out", 124, "HOSTED_VALIDATION_TIMED_OUT"),
+    ],
+)
+async def test_hosted_validation_fails_closed_when_terminal_failure_has_no_commands(
+    tmp_path: Path,
+    state: str,
+    expected_returncode: int,
+    expected_reason_code: str,
+) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": state,
+                    "message": "host-side job did not produce command results",
+                    "returncode": 0,
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+        )
+
+    assert not result.all_passed
+    assert len(result.commands) == 1
+    command = result.commands[0]
+    assert command.command == "hosted validation operation"
+    assert command.phase == "validate"
+    assert command.returncode == expected_returncode
+    assert command.reason_code == expected_reason_code
+    assert command.stderr_path.read_text(encoding="utf-8") == (
+        "host-side job did not produce command results\n"
+    )
+
+
+@pytest.mark.unit
+async def test_hosted_validation_rejects_command_output_over_max_output_bytes(
+    tmp_path: Path,
+) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "failed",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 1,
+                            "duration_seconds": 0.2,
+                            "stdout": "12345",
+                            "stderr": "",
+                            "phase": "validate",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(max_output_bytes=4),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match="output exceeds"):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
+
+    assert not (tmp_path / "ws_hosted" / "01_validate.stdout").exists()
+    assert not (tmp_path / "ws_hosted" / "01_validate.stderr").exists()
+
+
+@pytest.mark.unit
+async def test_hosted_validation_poll_allows_per_command_output_budget(
+    tmp_path: Path,
+) -> None:
+    output = "x" * 24_000
+    profile = WorkspaceProfile.model_validate(
+        {
+            "name": "hosted-test",
+            "phases": {
+                "validate": [
+                    "test command 1",
+                    "test command 2",
+                    "test command 3",
+                    "test command 4",
+                ]
+            },
+        }
+    )
+
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": f"test command {index}",
+                            "returncode": 0,
+                            "duration_seconds": 0.2,
+                            "stdout": output,
+                            "stderr": "",
+                            "phase": "validate",
+                        }
+                        for index in range(1, 5)
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(max_output_bytes=25_000),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=profile,
+            phase_names=("validate",),
+            include_coverage=False,
+        )
+
+    assert result.all_passed
+    assert [command.command for command in result.commands] == [
+        "test command 1",
+        "test command 2",
+        "test command 3",
+        "test command 4",
+    ]
+    assert (tmp_path / "ws_hosted" / "04_validate.stdout").read_text(encoding="utf-8") == output
+
+
+@pytest.mark.unit
+async def test_hosted_validation_terminal_failure_preserves_host_stderr(
+    tmp_path: Path,
+) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "failed",
+                    "stdout": "host stdout\n",
+                    "stderr": "host stderr\n",
+                    "command": "hosted validate",
+                    "phase": "post_agent",
+                    "reason_code": "HOSTED_CUSTOM_FAILURE",
+                    "duration_seconds": 2.5,
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+        )
+
+    command = result.commands[0]
+    assert command.command == "hosted validate"
+    assert command.phase == "post_agent"
+    assert command.reason_code == "HOSTED_CUSTOM_FAILURE"
+    assert command.duration_seconds == 2.5
+    assert command.stdout_path.read_text(encoding="utf-8") == "host stdout\n"
+    assert command.stderr_path.read_text(encoding="utf-8") == "host stderr\n"
+
+
+@pytest.mark.unit
+async def test_hosted_validation_rejects_missing_commands_payload(tmp_path: Path) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match="missing commands"):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
+
+
+@pytest.mark.unit
+async def test_hosted_validation_rejects_malformed_commands_payload(tmp_path: Path) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": {"command": "pytest -q"},
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match="malformed commands"):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
+
+
+@pytest.mark.unit
+async def test_hosted_validation_rejects_malformed_command_result(tmp_path: Path) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": ["not-a-command-object"],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match="command result is malformed"):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
+
+
+@pytest.mark.unit
+async def test_hosted_validation_normalizes_malformed_optional_command_maps(
+    tmp_path: Path,
+) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 0,
+                            "duration_seconds": 0.2,
+                            "stdout": "passed\n",
+                            "stderr": "",
+                            "phase": "validate",
+                            "stream_ids": "not-a-map",
+                            "metadata": "not-a-map",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+        )
+
+    assert result.commands[0].stream_ids == {}
+    assert result.commands[0].metadata == {}
+
+
+@pytest.mark.unit
+async def test_hosted_validation_accepts_valid_retry_count(tmp_path: Path) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 0,
+                            "duration_seconds": 0.2,
+                            "stdout": "",
+                            "stderr": "",
+                            "phase": "validate",
+                            "retry_count": 2,
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        result = await delegate.run_profile_phases(
+            workspace_id="ws_hosted",
+            compose_project="unused",
+            compose_file=tmp_path / "missing-compose.yml",
+            profile=WorkspaceProfile(name="hosted-test"),
+            phase_names=("validate",),
+        )
+
+    assert result.commands[0].retry_count == 2
+
+
+@pytest.mark.unit
+async def test_hosted_validation_rejects_invalid_retry_count(tmp_path: Path) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 0,
+                            "duration_seconds": 0.2,
+                            "stdout": "",
+                            "stderr": "",
+                            "phase": "validate",
+                            "retry_count": "one",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match="invalid integer field"):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
+
+
+@pytest.mark.unit
+async def test_hosted_validation_rejects_invalid_command_duration(tmp_path: Path) -> None:
+    async def _handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "POST" and request.url.path == "/v1/validation-runs":
+            return httpx.Response(
+                202,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "operation_url": "/v1/operations/val_1",
+                },
+            )
+        if request.method == "GET" and request.url.path == "/v1/operations/val_1":
+            return httpx.Response(
+                200,
+                json={
+                    "operation_id": "val_1",
+                    "workspace_id": "ws_hosted",
+                    "state": "succeeded",
+                    "commands": [
+                        {
+                            "command": "pytest -q",
+                            "returncode": 0,
+                            "duration_seconds": "0.2",
+                            "stdout": "",
+                            "stderr": "",
+                            "phase": "validate",
+                        }
+                    ],
+                },
+            )
+        raise AssertionError(f"unexpected request {request.method} {request.url}")
+
+    async with httpx.AsyncClient(transport=httpx.MockTransport(_handler)) as client:
+        delegate = HostedValidationDelegate(
+            _config(),
+            artifacts_dir=tmp_path,
+            client=client,
+        )
+        with pytest.raises(HostedDelegationProtocolError, match="missing duration_seconds"):
+            await delegate.run_profile_phases(
+                workspace_id="ws_hosted",
+                compose_project="unused",
+                compose_file=tmp_path / "missing-compose.yml",
+                profile=WorkspaceProfile(name="hosted-test"),
+                phase_names=("validate",),
+            )
