+++ tests/unit/service/test_pr_monitor_adoption_parts/test_pr_monitor_adoption_part_001.py
@@ -1471,59 +1471,3 @@ class TestPullRequestMonitorAdoptionServicePart001:
             assert fresh_task.external_id == fresh_workspace.task_external_id
             assert fresh_task.idempotency_key == f"{logical_key}:g2"
             assert fresh_task.title == "feature: retitled again"
-
-    @pytest.mark.unit
-    async def test_terminal_prior_reuses_canonical_key_despite_unrelated_generated_key(
-        self,
-        factory: async_sessionmaker[AsyncSession],
-    ) -> None:
-        logical_key = adoption_module.pr_adoption_idempotency_key(
-            repo_slug="dimileeh/aira-web",
-            pr_number=277,
-        )
-
-        async with factory() as session:
-            first = await PullRequestMonitorAdoptionService(
-                session,
-                metadata_fetcher=_MetadataFetcher(_metadata()),
-            ).adopt(
-                PullRequestMonitorAdoptionRequest(
-                    repo_slug="dimileeh/aira-web",
-                    pr_number=277,
-                )
-            )
-            old_workspace = await WorkspaceRepository(session).get(first.workspace_id)
-            assert old_workspace is not None
-            old_workspace.status = WorkspaceStatus.destroyed.value
-            await WorkspaceRepository(session).create(
-                repo_url="git@github.com:dimileeh/unrelated.git",
-                branch_base="main",
-                task_title="unrelated stale adoption key",
-                task_prompt="stale key collision fixture",
-                task_external_id="unrelated-stale-adoption-key",
-                agent="codex",
-                test_commands=[],
-                idempotency_key=f"{logical_key}:g1",
-            )
-            await session.commit()
-
-        async with factory() as session:
-            result = await PullRequestMonitorAdoptionService(
-                session,
-                metadata_fetcher=_MetadataFetcher(_metadata()),
-            ).adopt(
-                PullRequestMonitorAdoptionRequest(
-                    repo_slug="dimileeh/aira-web",
-                    pr_number=277,
-                )
-            )
-            await session.commit()
-
-        assert result.attached_existing is False
-        assert result.workspace_id != first.workspace_id
-
-        async with factory() as session:
-            fresh_workspace = await WorkspaceRepository(session).get(result.workspace_id)
-
-        assert fresh_workspace is not None
-        assert fresh_workspace.idempotency_key == logical_key
